Move password's controllers to a single file
